isolate prerender function
must be separate from core imports to save the selected activity menu
