+++ src/_reactComponents/Activity/SettingComponents.jsx
@@ -23,8 +23,7 @@ import ActionButtonGroup from '../PanelHeaderComponents/ActionButtonGroup';
 import ActionButton from '../PanelHeaderComponents/ActionButton';
 import { toastType, useToast } from '@Toast';
 import { searchParamAtomFamily } from '../../Tools/_framework/NewToolRoot';
-import { useSaveDraft } from '../../Tools/_framework/ToolPanels/DoenetMLEditor';
-import { prerenderActivity } from '../../_utils/activityUtils';
+import { prerenderActivity } from '../../_utils/activtyWebWorker';
 import Textfield from '../PanelHeaderComponents/Textfield';
 
 const InputWrapper = styled.div`
